requirements.txt
streamlit==1.36.0
supabase==2.7.4
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2